[#14] Small folder refactor
