+++ backend/src/main/java/c/team/participants/list/ParticipantListResponse.java
@@ -1,0 +1,12 @@
+package c.team.participants.list;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class ParticipantListResponse {
+    String sessionId;
+    String passcode;
+    String approvalRoomId;
+}
